cover mimetype verification

+++ utils/trackTester.py
@@ -85,6 +85,7 @@ class TrackTester:
         self._testMissorderedTags()
         # ErrorCode 19 : Cover is invalid (not 1000x1000 jpg/png)
         # ErrorCode 20 : Track has no cover
+        # ErrorCode 22 : Cover format is not optimized (not jpg)
         self._testCoverValidity()
 
 
@@ -205,13 +206,17 @@ class TrackTester:
             self.errorCounter += 1
             self.errors.append(ErrorEnum.MISSING_COVER)
         else:
-            with open('tmp.jpg', 'wb') as img: # Tmp extraction
-                img.write(self.track.cover)
-                omg = Image.open('tmp.jpg')
-                if omg.size[0] != 1000 and omg.size[1] != 1000:
-                    self.errorCounter += 1
-                    self.errors.append(ErrorEnum.INVALID_COVER)
-            os.remove('tmp.jpg') # GC
+            if self.track.coverType == 'image/png':
+                self.errorCounter += 1
+                self.errors.append(ErrorEnum.UNOPTIMAL_COVER)
+            else:
+                with open('tmp.jpg', 'wb') as img: # Tmp extraction
+                    img.write(self.track.cover)
+                    omg = Image.open('tmp.jpg')
+                    if omg.size[0] != 1000 and omg.size[1] != 1000:
+                        self.errorCounter += 1
+                        self.errors.append(ErrorEnum.INVALID_COVER)
+                os.remove('tmp.jpg') # GC
 
 
     # Tests a Track on a given topic using an error code as documented in this function
